refactor(i18n): consolidate translation tests into a single comprehensive test suite
